Renamed handler package, renamed Type to CallType
