+++ tygron-environment/src/main/java/nl/tudelft/contextproject/tygron/handlers/BooleanResultHandler.java
@@ -1,4 +1,4 @@
-package nl.tudelft.contextproject.tygron.results;
+package nl.tudelft.contextproject.tygron.handlers;
 
 public class BooleanResultHandler extends ResultHandler<Boolean> {
   @Override
