+++ tygron-environment/src/main/java/nl/tudelft/contextproject/tygron/handlers/JsonArrayResultHandler.java
@@ -1,4 +1,4 @@
-package nl.tudelft.contextproject.tygron.results;
+package nl.tudelft.contextproject.tygron.handlers;
 
 import org.json.JSONArray;
