+++ tygron-environment/src/main/java/nl/tudelft/contextproject/tygron/handlers/ResultHandler.java
@@ -1,4 +1,4 @@
-package nl.tudelft.contextproject.tygron.results;
+package nl.tudelft.contextproject.tygron.handlers;
 
 public abstract class ResultHandler<T> {
   public abstract T handleResult(String input);
